Use a single KMS client with reduced scope, add user-agent
This moves the creation of the KMS client into main where it can be shared with both the initialize() and unseal() functions. It also reduces the scope from cloud-platform to https://www.googleapis.com/auth/cloudkms.

This commit also adds a custom user-agent string to both the KMS and Storage clients.